+++ src/CoreServiceProvider.php
@@ -3,6 +3,7 @@
 namespace LaravelCore;
 
 use Illuminate\Support\ServiceProvider;
+use LaravelCore\Console\StarterKitInstallCommand;
 use LaravelCore\Menu\MenuRegistry;
 use LaravelCore\Settings\SettingsRegistry;
 
@@ -12,5 +13,14 @@ class CoreServiceProvider extends ServiceProvider
     {
         $this->app->singleton(MenuRegistry::class);
         $this->app->singleton(SettingsRegistry::class);
+    }
+
+    public function boot(): void
+    {
+        if ($this->app->runningInConsole()) {
+            $this->commands([
+                StarterKitInstallCommand::class,
+            ]);
+        }
     }
 }